ai-os/knowledge-pipeline/requirements.txt
anthropic>=0.30.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
networkx>=3.3
watchdog>=4.0.0
